scripts/process_data.py
#!/usr/bin/env python3
"""
Script para procesar y preparar datos para análisis.
"""

import geopandas as gpd
import pandas as pd
from pathlib import Path
from sqlalchemy import create_engine
import logging
import os
from dotenv import load_dotenv

# Cargar variables de entorno
load_dotenv()

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class DataProcessor:
    """Procesa y prepara datos para análisis."""

    def __init__(self):
        self.engine = self.create_db_connection()

    def create_db_connection(self):
        """Crea conexión a PostGIS."""
        db_url = (
            f"postgresql://{os.getenv('POSTGRES_USER')}:"
            f"{os.getenv('POSTGRES_PASSWORD')}@"
            f"{os.getenv('POSTGRES_HOST', 'localhost')}:"
            f"{os.getenv('POSTGRES_PORT', '5432')}/"
            f"{os.getenv('POSTGRES_DB')}"
        )
        return create_engine(db_url)

    def load_to_postgis(self, gdf, table_name, schema='raw_data'):
        """Carga GeoDataFrame a PostGIS."""
        try:
            gdf.to_postgis(
                table_name,
                self.engine,
                schema=schema,
                if_exists='replace',
                index=False
            )
            logger.info(f"Tabla {schema}.{table_name} creada exitosamente")
            return True
        except Exception as e:
            logger.error(f"Error cargando a PostGIS: {e}")
            return False

    def process_osm_network(self, input_file):
        """Procesa red vial de OSM."""
        logger.info("Procesando red vial...")
        # Implementar procesamiento
        pass

    def create_spatial_indices(self):
        """Crea índices espaciales en las tablas."""
        logger.info("Creando índices espaciales...")
        # Implementar creación de índices
        pass


def main():
    processor = DataProcessor()
    # Implementar lógica principal
    logger.info("Procesamiento completado!")


if __name__ == '__main__':
    main()
