requirements.txt
# Core Geospatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
rasterio>=1.3.0
fiona>=1.9.0
osmnx>=1.8.0

# Data Science
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Spatial Analysis
pysal>=23.0
esda>=2.5.0
splot>=1.1.5

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
folium>=0.15.0
streamlit>=1.29.0
streamlit-folium>=0.17.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0

# Remote Sensing / STAC
pystac-client>=0.7.5
planetary-computer>=1.0.0

# Archive extraction
rarfile>=4.0
